Agents/KickCount-Agent/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
langchain
langchain-groq
langchain-community
python-jose
langchain-pinecone
openai
pinecone